% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPM_AIC_CV_50Split.R
\name{LPM_AIC_CV_50Split}
\alias{LPM_AIC_CV_50Split}
\title{Recommended method for estimating $\sigma^2$}
\usage{
LPM_AIC_CV_50Split(XMat, yVec)
}
\description{
Recommended method for estimating $\sigma^2$
}
