% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIL.R
\name{MAIL}
\alias{MAIL}
\title{MAIL}
\usage{
MAIL(
  XMat,
  yVec,
  splitOption,
  firstSOILWeightType,
  smallestModelWeightType,
  firstSOILPsi,
  smallestModelPsi,
  sigma2EstFunc,
  trueSD = NULL,
  verbose = FALSE
)
}
\description{
\code{MAIL} runs the Model-Averaged Inferential Learning method under different parameter settings
}
\details{
The most important choice is whether or not use data splitting. 
The advantage of data splitting is to mitigate post selection changes to inference.
The advantage of using all of the data is to reduce bias.
There are six total parameters:
\itemize{
  \item splitOption \in \{"Full","Split"\}
}
}
