% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIL_Full.R
\name{MAIL_Full}
\alias{MAIL_Full}
\title{MAIL_Full}
\usage{
MAIL_Full(XMat, yVec)
}
\arguments{
\item{XMat}{a n by p numeric matrix}

\item{yVec}{a n by 1 numeric vector}
}
\value{
resList a list with the following elements:
}
\description{
\code{MAIL_Full} runs MAIL without data splitting
}
\details{
This is a specific use of the \code{\link{MAIL}} function.
The function uses the following arguments with MAIL:
\itemize{
  \item splitOption = "Full"
  \item firstSOILWeightType = "BIC"
  \item smallestModelWeightType = "AIC"
  \item firstSOILPsi = 0.5
  \item smallestModelPsi = 0
  \item sigma2EstFunc = "LPM_AIC_CV_50Split"
  \item verbose = FALSE
}
}
\examples{

two main ideas to explain with an example
\itemize {
  \item how to run MAIL_Full
  \item how to run diagnostics for MAIL_Full - is it a reliable method for the current data
}


#### how to run MAIL_Full

# first step - organize the data
# gene expression problem: response is the expression of gene probe X249.
# install package colonCA through Bioconductor (use these steps)
# https://bioconductor.org/packages/release/data/experiment/html/colonCA.html
require(colonCA)
data(colonCA)
dataSet <- t(colonCA@assayData$exprs)
colon_y <- dataSet[,249]
colon_x <- dataSet[,-1*249]
colon_x <- t(unique(t(colon_x))) # remove duplicate columns - original has p = 1999, reduce to p = 1990
colon_y <- scale(colon_y)
colon_x <- scale(colon_x)

colonMAILFull <- MAIL_Full(XMat = colon_x,yVec = colon_y)


# compare the confidence intervals for MAIL_Full and MAIL_Split
# for X257
}
\seealso{
\code{\link{MAIL}} and \code{\link{MAIL_Split}}
}
