% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIL_Split.R
\name{MAIL_Split}
\alias{MAIL_Split}
\title{MAIL_Split}
\usage{
MAIL_Split(XMat, yVec)
}
\arguments{
\item{XMat}{a matrix}

\item{yVec}{a vector}
}
\value{
resList a list
}
\description{
\code{MAIL_Split} runs MAIL with data splitting
}
\details{
This is a specific use of the \code{\link{MAIL}} function. 
The function uses the following arguments with MAIL: 
\itemize{
  \item splitOption = "Split"
  \item firstSOILWeightType = "BIC"
  \item smallestModelWeightType = "AIC"
  \item firstSOILPsi = 0.5
  \item smallestModelPsi = 0
  sigma2EstFunc = "LPM_AIC_CV_50Split"
  verbose = FALSE
}
}
