% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIL_Full.R
\name{MAIL_Full}
\alias{MAIL_Full}
\title{MAIL_Full}
\usage{
MAIL_Full(XMat, yVec)
}
\arguments{
\item{XMat}{a n by p numeric matrix}

\item{yVec}{a n by 1 numeric vector}
}
\value{
resList a list with the following elements:
\itemize{
  \item selectedSet: The indices of the selected variables (out of {1,...,p}).
  \item tempCI: A matrix (numSelected by 2) - each row corresponds to the 95% confidence interval for the full target of each selected variable (in the same order as selected set).
  \item margVar: A vector (length numSelected) of the sampling variances.
  \item betaHat: The vector of model-averaged estimates.
  \item estSigma2: The estimated residual variance.
  \item candMat: An indicator matrix of candidate models (numModels by p).
  \item origSelectedSet: The variables selected the initial sweep by SOIL.
}
}
\description{
\code{MAIL_Full} runs MAIL without data splitting
}
\details{
This is a specific use of the \code{\link{MAIL}} function.
The function uses the following arguments with MAIL:
\itemize{
  \item splitOption = "Full"
  \item firstSOILWeightType = "BIC"
  \item smallestModelWeightType = "AIC"
  \item firstSOILPsi = 0.5
  \item smallestModelPsi = 0
  sigma2EstFunc = "LPM_AIC_CV_50Split"
  verbose = FALSE
}
}
\examples{


#### some comment here
require(colonCA)
dataSet = t(colonCA@assayData$exprs)
colon_y = dataSet[,249]
colon_x = dataSet[,-1*249]
colon_x = t(unique(t(colon_x)))
colon_y = scale(colon_y)
}
\seealso{
\code{\link{MAIL}} and \code{\link{MAIL_Split}}
}
