% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIL_Full.R
\name{MAIL_Full}
\alias{MAIL_Full}
\title{MAIL_Full}
\usage{
MAIL_Full(XMat, yVec)
}
\arguments{
\item{XMat}{a n by p numeric matrix}

\item{yVec}{a n by 1 numeric vector}
}
\value{
resList a list
}
\description{
\code{MAIL_Full} runs MAIL without data splitting
}
\details{
This is a specific use of the \code{\link{MAIL}} function. 
The function uses the following arguments with MAIL: 
\itemize{
  \item splitOption = "Full"
  \item firstSOILWeightType = "BIC"
  \item smallestModelWeightType = "AIC"
  \item firstSOILPsi = 0.5
  \item smallestModelPsi = 0
  sigma2EstFunc = "LPM_AIC_CV_50Split"
  verbose = FALSE
}
}
