% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_MAIL_Functions.R
\name{MAIL_Full}
\alias{MAIL_Full}
\title{Both MAIL_Full and MAIL_Split use the same basic algorithm.
MAIL_Full and MAIL_Split have preset parameters that control whether or not to split the data, the weight types, the complexity penalties and the method for estimating the error variance.
This basic algorithm is contained in the MAIL function, and this was found to have good coverage under a range of parameter settings.
Other splittings besides 50/50 were tried, and we found good coverage with n/4 or n/logn for the exploratory share in many cases, but using n/8 often resulted in undercoverage.
type something here}
\usage{
MAIL_Full(XMat, yVec)
}
\description{
Both MAIL_Full and MAIL_Split use the same basic algorithm.
MAIL_Full and MAIL_Split have preset parameters that control whether or not to split the data, the weight types, the complexity penalties and the method for estimating the error variance.
This basic algorithm is contained in the MAIL function, and this was found to have good coverage under a range of parameter settings.
Other splittings besides 50/50 were tried, and we found good coverage with n/4 or n/logn for the exploratory share in many cases, but using n/8 often resulted in undercoverage.
type something here
}
